#4 Add basic setup of webapp-api backend service
Creates the very basic, including dependencies, spring boot application for the webapp-api backend service.

Due to the nature of the monorepo and the target of having two backends, I've already moved the `gradle` setup around so that it's easy to add the next project. The advantages of this are many, i.e. we can in the future combine frontend + backend builds into a single gradle command, but can also manage dependencies of the java services at a single point.

Signed-off-by: RicardoMonteiroSimoes <30594398+RicardoMonteiroSimoes@users.noreply.github.com>